% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Cort.R
\name{biv_rho}
\alias{biv_rho}
\alias{biv_rho,Cort-method}
\title{Spearman's rho matrix of a copula}
\usage{
biv_rho(copula)

\S4method{biv_rho}{Cort}(copula)
}
\arguments{
\item{copula}{the copula object}
}
\value{
the density of the copula on each observation
}
\description{
Computes the bivariate Spearmann's rho matrix for a copula.
}
\section{Functions}{
\itemize{
\item \code{biv_rho,Cort-method}: Method for the class Cort
}}

\examples{
cop <- Cort(LifeCycleSavings[,1:3])
biv_rho(cop)

}
