% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Cort.R
\name{biv_tau}
\alias{biv_tau}
\alias{biv_tau,Cort-method}
\title{Kendall's tau matrix of a copula}
\usage{
biv_tau(copula)

\S4method{biv_tau}{Cort}(copula)
}
\arguments{
\item{copula}{the copula object}
}
\value{
the density of the copula on each observation
}
\description{
Computes the bivariate Kendall's tau matrix for a copula.
}
\section{Functions}{
\itemize{
\item \code{biv_tau,Cort-method}: Method for the class Cort
}}

\examples{
cop <- Cort(cort::funcdep_data[1:10,1:3])
biv_tau(cop)

}
