% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box.R
\name{Box-Class}
\alias{Box-Class}
\alias{Box}
\alias{show,Box-method}
\title{A simple Box}
\usage{
Box(a, b)

\S4method{show}{Box}(object)
}
\arguments{
\item{a}{a vector for the minimum of the box}

\item{b}{a vector for the maximum of the box}

\item{object}{The box}
}
\value{
a Box object
}
\description{
Box class
}
\details{
The Box class represent a box object with a dimension, a minimum and a maximum point
This class is useful to produce later partitioning trees that are needed
for the CORT algorithm.
}
\section{Functions}{
\itemize{
\item \code{show,Box-method}: Prints simple dispatch of the box
}}

\examples{
library(cort)
b = Box(rep(1/4,4),rep(3/4,4))
contains(b,rep(1/2,4))
}
