% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Cort.R, R/CortForest.R
\name{constraint_infl}
\alias{constraint_infl}
\alias{constraint_infl,Cort-method}
\alias{constraint_infl,CortForest-method}
\title{Constraint influence of the model}
\usage{
constraint_infl(object)

\S4method{constraint_infl}{Cort}(object)

\S4method{constraint_infl}{CortForest}(object)
}
\arguments{
\item{object}{the copula object}
}
\value{
The constraint influence statistic of the model
}
\description{
Compute the constraint influence of the model
}
\section{Functions}{
\itemize{
\item \code{constraint_infl,Cort-method}: Method for the class Cort

\item \code{constraint_infl,CortForest-method}: Method for the class Cort
}}

\examples{
cop <- Cort(LifeCycleSavings[,1:3])
constraint_infl(cop)

}
