% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Box.R
\name{contains}
\alias{contains}
\alias{contains,Box,matrix-method}
\title{Contains method}
\usage{
contains(object, u, type = "loose")

\S4method{contains}{Box,matrix}(object, u, type = "loose")
}
\arguments{
\item{object}{an object of class \code{Box}}

\item{u}{a matrix object, representing points of the right dimension in each row}

\item{type}{a string, one of c("strict","loose","rllc") (default : loose)}
}
\value{
A vector of boolean values, one for each row of the data \code{u}
}
\description{
Answer the question : "Does the box contains those points ? The \code{type} parameter is used to say if you want the test to be strict on both side, loose on both side, or right-limited and left-continuous via the \code{rllc} possibility.
}
\section{Functions}{
\itemize{
\item \code{contains,Box,matrix-method}: Check if points are contained in the box
}}

\examples{
library(cort)
contains(Box(rep(0,2),rep(1/2,2)),rep(2,2))
}
