% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvexCombCopula.R
\name{ConvexCombCopula-class}
\alias{ConvexCombCopula-class}
\alias{ConvexCombCopula}
\alias{dim,ConvexCombCopula-method}
\title{Convex Combination of copulas.}
\usage{
ConvexCombCopula(copulas, alpha = rep(1, length(copulas)))

\S4method{dim}{ConvexCombCopula}(x)
}
\arguments{
\item{copulas}{a list of copulas of same dimention}

\item{alpha}{a vector of (positive) weights}

\item{x}{ConvexCombCopula object}
}
\value{
a ConvexCombCopula object
}
\description{
ConvexCombCopula class
}
\details{
The ConvexCombcopula class is used to build convex combinations of copulas,
with given positives weights. The rCopula and pCopula functions works for
thoose copulas, assuming they work for the given copulas that we combined
in a convex way.
}
\section{Methods (by generic)}{
\itemize{
\item \code{dim}: dimension
}}

\examples{
library(empCop)
copulas <- list(
  copula::archmCopula('gumbel',3,dim=2),
  copula::archmCopula('clayton',-1,dim=2)
)

alpha <- c(1,4)

(cop <- ConvexCombCopula(copulas,alpha))
}
