% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cort.R, R/CortForest.R, R/empiricalCopula.R
\name{Cort-Class}
\alias{Cort-Class}
\alias{Cort}
\alias{dim,Cort-method}
\alias{CortForest}
\alias{dim,empiricalCopula-method}
\title{The Cort estimator}
\usage{
Cort(
  x,
  p_value_for_dim_red = 0.75,
  min_node_size = 1,
  pseudo_data = FALSE,
  number_max_dim = NULL,
  verbose_lvl = 1
)

\S4method{dim}{Cort}(x)

CortForest(
  x,
  p_value_for_dim_red = 0.75,
  n_trees = 10,
  compte_loo_weights = FALSE,
  min_node_size = 1,
  pseudo_data = FALSE,
  number_max_dim = NULL,
  verbose_lvl = 2
)

\S4method{dim}{empiricalCopula}(x)
}
\arguments{
\item{x}{The data, must be provided as a matrix with each row as an observation.}

\item{p_value_for_dim_red}{a p_value for the localised dimension reduction test}

\item{min_node_size}{The minimum number of observation avaliable in a leaf to initialise a split.}

\item{pseudo_data}{set to True if you are already providing data on the copula space.}

\item{number_max_dim}{The maximum number of dimension a split occurs in. Defaults to be all of the dimensions.}

\item{verbose_lvl}{verbosity level : can be 0 (none) or an integer. bigger the interger bigger the ouput level.}

\item{n_trees}{Number of trees}

\item{compte_loo_weights}{Defaults to FALSE. Allows to use an automatic re-weighting of the trees in the forest, based on leave-one-out considerations.}
}
\value{
a Cort object that can be fitted easily to produce a copula estimate.

a Cort object that can be fitted easily to produce a copula estimate.
}
\description{
Cort class

CortForest class
}
\details{
This class implements the CORT algorithm to a fit a multivariate copula using piece constant density.

This class implements the bagging of CORT models, with an oob error minimisation in the weights.
}
\section{Functions}{
\itemize{
\item \code{dim,Cort-method}: dimension of the copula

\item \code{dim,empiricalCopula-method}: dimension of the copula
}}

\examples{
library(cort)
(Cort(LifeCycleSavings[,1:3]))
library(cort)
(CortForest(LifeCycleSavings[,1:3]))
}
