% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalCopula.R
\docType{class}
\name{empiricalCopula-class}
\alias{empiricalCopula-class}
\alias{dim,empiricalCopula-method}
\title{Empirical Copula class (virtual mother class)}
\usage{
\S4method{dim}{empiricalCopula}(x)
}
\arguments{
\item{x}{ConvexCombCopula object}
}
\description{
Empirical Copula class (virtual mother class)
}
\section{Methods (by generic)}{
\itemize{
\item \code{dim}: dimension
}}

\section{Slots}{

\describe{
\item{\code{pseudo_data}}{matrix : pseudo_data that the empirical copula is based on.}
}}

