% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Box.R
\name{intersect}
\alias{intersect}
\alias{intersect,Box,Box-method}
\title{Intersection of 2 boxes}
\usage{
intersect(object, b)

\S4method{intersect}{Box,Box}(object, b)
}
\arguments{
\item{object}{an object of class \code{Box}}

\item{b}{an other object of class \code{Box}}
}
\value{
A box object representing the intersection of those two boxes.
}
\description{
Gives the intersection of two boxes
}
\section{Functions}{
\itemize{
\item \code{intersect,Box,Box-method}: Compute the intersection of 2 boxes
}}

\examples{
library(cort)
intersect(Box(rep(0,2),rep(1/2,2)),Box(rep(1/4,2),rep(1/2,2)))
}
