% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is_splittable}
\alias{is_splittable}
\title{Check if a box is splittable}
\usage{
is_splittable(object)
}
\arguments{
\item{object}{an object of class \code{Box}}
}
\value{
a boolean telling you if the box is splittable or not
}
\description{
To be splitted, a box needs to have at least min_node_size points and 2 splitting dimensions.
}
\examples{
library(cort)
b = Box(rep(0,2),rep(1,2))
b = WeightedBox(b,matrix(0.5,nrow=2,ncol=2),1,c(1,2))
is_splittable(b)
}
