% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Box.R
\name{measure_in}
\alias{measure_in}
\alias{measure_in,Box,matrix-method}
\title{Measure of a point inside a box}
\usage{
measure_in(object, u)

\S4method{measure_in}{Box,matrix}(object, u)
}
\arguments{
\item{object}{an object of class \code{Box}}

\item{u}{a matrix object, representing points of the right dimension in each row}
}
\value{
A vector of values between 0 and 1, representing the measure of each point inside the box.
}
\description{
compute the formula lambda(0,u inter b) / lambda (b) for a box b and a point u
}
\section{Functions}{
\itemize{
\item \code{measure_in,Box,matrix-method}: Measure points in the box
}}

\examples{
library(cort)
measure_in(Box(rep(0,2),rep(1/2,2)),matrix(runif(100),nrow=50,ncol=2))
}
