% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/cbCopula.R, R/cbkmCopula.R
\name{pCopula}
\alias{pCopula}
\alias{pCopula,matrix,cbCopula-method}
\alias{pCopula,matrix,cbkmCopula-method}
\alias{dCopula,matrix,cbkmCopula-method}
\title{Copula density}
\usage{
pCopula(u, copula, ...)

\S4method{pCopula}{matrix,cbCopula}(u, copula)

\S4method{pCopula}{matrix,cbkmCopula}(u, copula)

\S4method{dCopula}{matrix,cbkmCopula}(u, copula)
}
\arguments{
\item{u}{numeric matrix : one row per observation}

\item{copula}{the copula object}

\item{...}{other parameter to be passed to methods for this generic.}
}
\value{
the density of the copula on each observation
}
\description{
This function returns the value of the copula itself on given points.
}
\section{Functions}{
\itemize{
\item \code{pCopula,matrix,cbCopula-method}: Method for the cbCopula

\item \code{pCopula,matrix,cbkmCopula-method}: Method for the cbCopula

\item \code{dCopula,matrix,cbkmCopula-method}: Method for the cbCopula
}}

\examples{
library(cort)
# The exmeple needs to be re-done.
}
