% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Box.R
\name{project}
\alias{project}
\alias{project,Box,numeric-method}
\title{project a box onto smaller dimensions}
\usage{
project(object, dimensions)

\S4method{project}{Box,numeric}(object, dimensions)
}
\arguments{
\item{object}{an object of class \code{Box}}

\item{dimensions}{the dimensions to project on}
}
\value{
a matrix with each random point in a row.
}
\description{
Given new dimensions, a box can be projected on these dimensions.
}
\section{Functions}{
\itemize{
\item \code{project,Box,numeric-method}: projects a box onto sub dimensions.
}}

\examples{
library(cort)
project(Box(rep(0,2),rep(1/2,2)),1)
}
