% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Cort.R
\name{project_on_dims}
\alias{project_on_dims}
\alias{project_on_dims,Cort-method}
\title{Projection on smaller dimensions of a copula (if implemented)}
\usage{
project_on_dims(object, dims)

\S4method{project_on_dims}{Cort}(object, dims)
}
\arguments{
\item{object}{: the tree}

\item{dims}{the set of dimensions}
}
\value{
other cort object
}
\description{
Currently only implemented for Cort models.
Compute, as a Cort object, the projection on a smaller set of dimensions of a Cort object.
}
\section{Functions}{
\itemize{
\item \code{project_on_dims,Cort-method}: Method for the class Cort
}}

\examples{
cop <- Cort(LifeCycleSavings[,1:3])
projection = project_on_dims(cop,c(1,2))

}
