% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Cort.R, R/CortForest.R
\name{quad_norm}
\alias{quad_norm}
\alias{quad_norm,Cort-method}
\alias{quad_norm,CortForest-method}
\title{Quadratic norm of the model}
\usage{
quad_norm(object)

\S4method{quad_norm}{Cort}(object)

\S4method{quad_norm}{CortForest}(object)
}
\arguments{
\item{object}{the copula object}
}
\value{
the Integrated square error quad_norm of the model
}
\description{
Compute the L2 norm of the model
}
\section{Functions}{
\itemize{
\item \code{quad_norm,Cort-method}: Method for the class Cort

\item \code{quad_norm,CortForest-method}: Method for the class Cort
}}

\examples{
cop <- Cort(LifeCycleSavings[,1:3])
quad_norm(cop)

}
