% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Cort.R
\name{quad_prod}
\alias{quad_prod}
\alias{quad_prod,Cort,Cort-method}
\title{Quadratic product of two copulas (if they have one)}
\usage{
quad_prod(object, other_tree)

\S4method{quad_prod}{Cort,Cort}(object, other_tree)
}
\arguments{
\item{object}{: the tree}

\item{other_tree}{: the other tree}
}
\value{
the quadratic product between the trees
}
\description{
Currently only implemented for Cort models.
Compute the L2 quadratic product of 2 trees
}
\section{Functions}{
\itemize{
\item \code{quad_prod,Cort,Cort-method}: Method for the class Cort
}}

\examples{
cop <- Cort(LifeCycleSavings[,1:3])
all.equal(quad_prod(cop,cop),quad_norm(cop))

}
