% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Cort.R
\name{quad_prod_with_data}
\alias{quad_prod_with_data}
\alias{quad_prod_with_data,Cort-method}
\title{Quadratic product with data of the model (if it has one)}
\usage{
quad_prod_with_data(object)

\S4method{quad_prod_with_data}{Cort}(object)
}
\arguments{
\item{object}{the copula object}
}
\value{
the quad_prod_with_data of the model
}
\description{
Currently only implemented for Cort models.
Compute the quadratic product with the empirical density from the data
}
\section{Functions}{
\itemize{
\item \code{quad_prod_with_data,Cort-method}: Method for the class Cort
}}

\examples{
cop <- Cort(LifeCycleSavings[,1:3])
quad_prod_with_data(cop)

}
