% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recoveryourself_data}
\alias{recoveryourself_data}
\title{Dataset recoveryourself_data}
\format{
A matrix with 500 rows and 2 columns

The example section below gives the code to re-generate this data if needed.
}
\usage{
recoveryourself_data
}
\description{
This dataset is a simple test: we simulate random samples from a density inside the piecewise copula class,
and test whether or not the estimator can recover it. For that, we will use a 2-dimensional sample with 500
observations, uniform on the unit hypercube, and apply the following function:
\deqn{h(u) = (u_1, \frac{u_2 + I_{u_1 \le \frac{1}{4}} + 2I_{u_1 \le \frac{1}{2}} + I_{\frac{3}{4} \le u_1}}{4})}
}
\details{
This dataset is studied in O. Laverny, V. Maume-Deschamps, E. Masiello and D. Rullière (2020).
}
\examples{
set.seed(seed = 12, kind = "Mersenne-Twister", normal.kind = "Inversion")
x = matrix(runif(1000),500,2)
recoveryourself_data = t(apply(x, 1,function(u){
  if(u[1]< 1/4){
    u[2] = 3/4 + u[2]/4
  } else{ if(u[1]<1/2){
    u[2] = 1/2 + u[2]/4
  } else { if(u[1]<3/4){
    u[2] = u[2]/4
  } else {
    u[2] = 1/4 + u[2]/4
  }}}
  return(u)
}))

}
\references{
\insertRef{laverny2020}{cort}
}
\keyword{datasets}
