% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Box.R
\name{simu_unif}
\alias{simu_unif}
\alias{simu_unif,Box,numeric-method}
\title{Simulate uniformly from a box}
\usage{
simu_unif(object, n)

\S4method{simu_unif}{Box,numeric}(object, n)
}
\arguments{
\item{object}{an object of class \code{Box}}

\item{n}{the number of points to simulate}
}
\value{
a matrix with each random point in a row.
}
\description{
simulate random points uniformly inside a box.
}
\section{Functions}{
\itemize{
\item \code{simu_unif,Box,numeric-method}: Simulate uniformly from the box
}}

\examples{
library(cort)
simu_unif(Box(rep(0,2),rep(1/2,2)),10)
}
