% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Box.R
\name{split}
\alias{split}
\alias{split,Box-method}
\title{Split a box onto a given breakpoint}
\usage{
split(object, ...)

\S4method{split}{Box}(object, breakpoint, breakpoint_dim)
}
\arguments{
\item{object}{an object of class \code{Box}}

\item{...}{other parameters that will be passed to methods.}

\item{breakpoint}{a point to split the box on, should be inside the box}

\item{breakpoint_dim}{the dimensions where you want the splitting to happened.}
}
\value{
a matrix with each random point in a row.
}
\description{
This functions splits a box onto given dimensions and given breakpoint
}
\section{Functions}{
\itemize{
\item \code{split,Box-method}: Splits a box on a given breakpoint and breakpoints dimensions.
}}

\examples{
library(cort)
split(Box(rep(0,2),rep(1/2,2)),rep(1/4,2),c(1,2))
}
