% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedBox.R
\name{WeightedBox-Class}
\alias{WeightedBox-Class}
\alias{WeightedBox}
\alias{show,WeightedBox-method}
\alias{is_splittable,WeightedBox-method}
\alias{split,WeightedBox-method}
\title{A simple Box}
\usage{
WeightedBox(box, data, weight = 1, split_dims = 1:box@dim, min_node_size = 1)

\S4method{show}{WeightedBox}(object)

\S4method{is_splittable}{WeightedBox}(object)

\S4method{split}{WeightedBox}(object, p_val_threshold = 0.75, number_max_dim = object@dim)
}
\arguments{
\item{box}{an object of class \code{Box}}

\item{data}{the data that the box contains.}

\item{weight}{(default = 1) a weight between 0 and 1.}

\item{split_dims}{(defaults : all)}

\item{min_node_size}{minimum number of data point in the box.}

\item{object}{The WeightedBox}

\item{p_val_threshold}{The threshold for the p_value of local dimension reduction tests}

\item{number_max_dim}{allow to perform optimization only on a subset of dimensions.}
}
\value{
a Box object
}
\description{
WeightedBox class
}
\details{
A WeightedBox is a box but containing more things : a weight, some data, and a split dimensions mainly.
To initialise it, you must provide a well-defined box. The splitting dimensions and the weights are optionals.
}
\section{Functions}{
\itemize{
\item \code{show,WeightedBox-method}: Prints simple dispatch of the box

\item \code{is_splittable,WeightedBox-method}: Check if the box can be splitted

\item \code{split,WeightedBox-method}: Splits the box and perform fitting of weights for the density part.
}}

\examples{
library(cort)
b = Box(rep(0,2),rep(1,2))
WeightedBox(b,matrix(0.5,nrow=2,ncol=2),1,c(1,2))
b = Box(rep(0,5),rep(1,5))
data = apply(LifeCycleSavings,2,rank)/(nrow(LifeCycleSavings)+1)
wb = WeightedBox(b,data,1,1:5)
new_boxes = split(wb)
}
