% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box.R
\name{ZeroBox}
\alias{ZeroBox}
\title{An empty box}
\usage{
ZeroBox(dim)
}
\arguments{
\item{dim}{the dimension of the box}
}
\value{
a Box object
}
\description{
ZeroBox constructor
}
\details{
This function construct a box with 0 volume.
}
\examples{
ZeroBox(4)
}
