% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbCopula.R
\name{cbCopula}
\alias{cbCopula}
\title{cbCopula contructor}
\usage{
cbCopula(x, m = rep(nrow(x), ncol(x)), pseudo = FALSE)
}
\arguments{
\item{x}{the data to be used}

\item{m}{checkerboard parameters}

\item{pseudo}{Boolean, defaults to \code{FALSE}. Set to \code{TRUE} if you are already
providing pseudo datas into the \code{x} argument.}
}
\value{
a cbCopula object
}
\description{
cbCopula contructor
}
\details{
The checkerboard copula is a kind of patchwork copula that only uses independant copula as fill-in, only where there are values on the empirical data provided. To create such a copula, you should provide data and checkerboard parameters (depending on the dimention of the data). To read more about thoose models, we refer to the work of Durante & Al, but also on the work of Cuberos & Al
}
