% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clayton_data}
\alias{clayton_data}
\title{Dataset clayton_data}
\format{A matrix with 200 rows and 4 columns}
\usage{
clayton_data
}
\description{
This dataset is a simulation of 200 points from a 3-dimensional clayton copula with \eqn{\theta = 7},
hence highly dependent, for the first, third and fourth marginals. The second marginal is added
as independent uniform draws. Lastly, the third marginal is flipped,
inducing a negative dependence structure.
}
\details{
This dataset is studied in O. Laverny, V. Maume-Deschamps, E. Masiello and D. Rullière (2020).
}
\references{
\insertRef{laverny2020}{cort}
}
\keyword{datasets}
