% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree.R
\name{Cort-Class}
\alias{Cort-Class}
\alias{Cort}
\alias{show,Cort-method}
\alias{fit,Cort-method}
\title{The Cort estimator}
\usage{
Cort(
  data,
  p_value_for_dim_red = 0.75,
  min_node_size = 1,
  pseudo_data = FALSE,
  number_max_dim = NULL,
  verbose = FALSE
)

\S4method{show}{Cort}(object)

\S4method{fit}{Cort}(object)
}
\arguments{
\item{data}{The data, must be provided as a matrix with each row as an observation.}

\item{p_value_for_dim_red}{a p_value for the localised dimension reduction test}

\item{min_node_size}{The minimum number of observation avaliable in a leaf to initialise a split.}

\item{pseudo_data}{set to True if you are already providing data on the copula space.}

\item{number_max_dim}{The maximum number of dimension a split occurs in. Defaults to be all of the dimensions.}

\item{verbose}{set to TRUE for more detailed output during the fit.}

\item{object}{The Cort object}
}
\value{
a Cort object that can be fitted easily to produce a copula estimate.
}
\description{
Cort class
}
\details{
This class implements the CORT algorithm to a fit a multivariate copula using piece constant density.
}
\section{Functions}{
\itemize{
\item \code{show,Cort-method}: Prints simple dispatch of the tree

\item \code{fit,Cort-method}: Fit the tree
}}

\examples{
library(cort)
Cort(LifeCycleSavings)
}
