% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{funcdep_data}
\alias{funcdep_data}
\title{Dataset funcdep_data}
\format{A matrix with 500 rows and 3 columns}
\usage{
funcdep_data
}
\description{
This dependence structure is constructed by applying the function :
\deqn{h(u_1,u_2,u_3) = (u_{1},\sin(2\pi u_{1})-\frac{u_{2}}{\pi},(1+\frac{u_{3}}{\pi^{2}})(\frac{u_{3}}{2} I_{\frac{1}{4}\ge u_1}-\sin(\pi^{x_{1}}) I_{\frac{1}{4} < u_{1}}))}
to uniformly drawn 3-dimensional random vectors. The dataset is the ranks of \eqn{h(u)}.
}
\details{
This dataset is studied in O. Laverny, V. Maume-Deschamps, E. Masiello and D. Rullière (2020).
}
\references{
\insertRef{laverny2020}{cort}
}
\keyword{datasets}
