% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Cort.R
\name{loss}
\alias{loss}
\alias{loss,Cort-method}
\title{Loss of a copula estimation (if the model has one)}
\usage{
loss(object)

\S4method{loss}{Cort}(object)
}
\arguments{
\item{object}{the copula object}
}
\value{
the Integrated square error loss of the model
}
\description{
Currently only implemented for Cort models.
Compute the loss of the model.
}
\section{Functions}{
\itemize{
\item \code{loss,Cort-method}: Method for the class Cort
}}

\examples{
cop <- Cort(cort::recoveryourself_data[1:10,])
loss(cop)

}
